// src/api/generated/types.gen.ts
export type TodoSchema = {
  /**
   * Id
   */
  id: string;
  /**
   * User
   */
  user: string;
  /**
   * Title
   */
  title: string;
  /**
   * Description
   */
  description: string;
  /**
   * Completed
   */
  completed: boolean;
  /**
   * Priority
   */
  priority: string;
  /**
   * Created At
   */
  created_at: string;
  /**
   * Updated At
   */
  updated_at: string;
};