// src/api/generated/zod.gen.ts
export const zCreateTodoSchema = z.object({
  title: z.string(),
  description: z.string().optional().default(''),
  completed: z.boolean().optional().default(false),
  priority: z.string().optional().default('medium'),
});